streamlit_app_files/requirements.txt
pandas
plotly_express
joblib
pydeck
requests
scikit-learn
catboost